'check' command added
The 'check' command will launch an immediate host or service check
depending on the active object.

Obviously we could prompt for a preferred date/time (the API supports
it), but I have yet to find anyone *not* using this call for immediate
checks. If you are that one (hi!) please send me a mail and fork this
repo from tehmaze so you can add patches.